refactor(github): split service.go into domain subfiles
service.go was 2,842 lines covering five unrelated domains. This is a
pure file-level split with no public API or behavior change:

- errors.go: ErrInvalidPRURL, ErrTaskNotFound, ErrSelfApprove +
  isTaskNotFound helper
- service.go: Service struct, NewService, getters/setters, interfaces
  (now 169 lines)
- service_tokens.go: GetStatus, ConfigureToken, ClearToken, rate-limit
  info
- service_pr.go: GetPR/GetPRStatus/GetPRFeedback/GetPRFiles/GetPRCommits,
  SubmitReview, MergePR, batch fetch, comparators
- service_pr_watch.go: PR watch CRUD, task-PR association, SyncTaskPR,
  TriggerPRSync*
- service_reviews.go: Review watch CRUD, fetch/enrich, repo/org/branch
  discovery, search, dedup row ops, GetPRStats
- service_issues.go: Issue watch CRUD + check + dedup ops
- service_cleanup.go: Review + issue task cleanup (shared failure
  tracking)
- service_pr_status.go: pure compute helpers

Each new file is under the 800-line revive threshold. The Service
struct, its constructor, and every public method signature are
preserved verbatim, so callers in cmd/kandev/* and
internal/orchestrator/event_handlers_github.go need no changes.

Three small linter fixes carried along (golangci runs with new: true,
so moving the funcs reflags pre-existing issues): rewrote
CheckPRWatch's hasNew assignment as a merged conditional, and added
narrow nolint directives for SyncTaskPR cyclomatic complexity (field-
by-field reconciler), UpdateIssueWatch cyclomatic complexity (mirrors
UpdateReviewWatch), and the pr_merged_or_closed reason string
(referenced by test).
